Add Character hit test to unit_test.cpp

+++ Character.cpp
@@ -96,7 +96,7 @@ Character Character::parseUnit(const std::string& name) {
 	std::map<std::string, std::string> values = parser.jsonParser(file);
 	if (values.find("name") != values.end() && values.find("hp") != values.end() && values.find("dmg") != values.end() && values.find("acd") != values.end()) {
 		file.close();
-		return Character(values["name"], stoi(values["hp"]), stoi(values["dmg"]), stoi(values["acd"]));
+		return Character(values["name"], stoi(values["hp"]), stoi(values["dmg"]), stod(values["acd"]));
 	}
 	else throw "incorrect values";
 }
